Update spring project

+++ BE/src/main/java/com/hong/ForPaw/controller/DTO/ShelterResponse.java
@@ -6,11 +6,19 @@ public class ShelterResponse {
 
     public record FindShelterListDTO(List<ShelterDTO> shelterDTOS){}
 
-    public record  FindShelterByIdDTO(String careAddr, String careTel, List<AnimalDTO> animals) {}
+    public record FindShelterByIdDTO(String careAddr,
+                                     String careTel,
+                                     List<AnimalDTO> animals) {}
 
-    public record AnimalDTO(Long id, String name, String age,
-                            String gender, String specialMark, String region,
-                            Integer inquiryNum, Integer likeNum, Boolean isLike,
+    public record AnimalDTO(Long id,
+                            String name,
+                            String age,
+                            String gender,
+                            String specialMark,
+                            String region,
+                            Long inquiryNum,
+                            Long likeNum,
+                            Boolean isLike,
                             String profileURL
     ){};
